refactor: clean up folder structure and naming
- extract config
- move all app-related handling to handlerUtil
- move purely http-related stuff to handlehttp
- move html-templates out of handlehttp to separate frontend
