+++ handlehttp/html-mapper.go
@@ -1,20 +1,17 @@
 package handlehttp
 
 import (
-	"embed"
 	"html/template"
+	"io/fs"
 	"log"
 	"net/http"
 )
-
-//go:embed html/*.gohtml
-var tplFS embed.FS
 
 func activateHtmlResponse(w http.ResponseWriter) {
 	w.Header().Set("content-type", Html.String())
 }
 
-func HtmlMapper(tplPaths ...string) ResponseMapper {
+func HtmlMapper(tplFS fs.FS, tplPaths ...string) ResponseMapper {
 	tpl := template.Must(template.ParseFS(tplFS, tplPaths...))
 
 	return func(w http.ResponseWriter, data any) {
