--- handlehttp/html/base.gohtml
@@ -1,11 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>test</title>
-</head>
-  <body>
-  test
-  </body>
-</html>
